requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
alembic==1.15.2
asyncpg==0.30.0
greenlet==3.2.2
psycopg2-binary==2.9.10
neo4j==5.28.1
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic[email]
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
celery==5.5.1
redis==5.2.1
boto3==1.38.3
openpyxl==3.1.5
requests==2.32.3
